# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Harmoniq App API",
    "description": "**Harmoniq App** is a backend API for a modern web application that allows users to register, authenticate, save favorite articles, and manage their profile. The API follows a `RESTful` structure and returns responses in `JSON format`.\n\n### Key Features:\n- User registration with avatar upload (multipart/form-data supported)\n- Login and logout functionality using JWT authentication\n- Access and Refresh token generation\n- Email verification after registration\n- File uploads to Cloudinary\n- Articles collection saves in MongoDB\n- Route protection using authentication middleware\n- Request validation using Joi\n\n### Technologies Used:\n- **Node.js** + **Express.js** — backend platform\n- **MongoDB** + **Mongoose** — database and ORM\n- **JWT (JSON Web Tokens)** — authentication system\n- **Multer** — file handling for multipart/form-data\n- **Cloudinary** — cloud image storage\n- **Joi** — input validation\n- **dotenv** — environment variable management\n- **Swagger (OpenAPI)** — API documentation\n\n### Source code:\n[https://github.com/StasSkv/harmoniq-app-backend](https://github.com/StasSkv/harmoniq-app-backend)\n",
    "version": "1.0.0",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/users/saved-articles": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get saved articles for the authenticated user",
        "operationId": "getSavedArticles",
        "description": "Returns a list of articles saved by the currently authenticated user.\n\nRequires a valid Bearer token to identify the user.\n\nThe saved articles are fetched by populating the `savedArticles` field in the user document.\n\n### Use cases:\n- Display user's saved articles on their dashboard\n- Allow users to manage saved reading content\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of saved articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Saved articles fetched successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Article"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/created-articles": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get articles created by the authenticated user (DEPRECATED)",
        "operationId": "getCreatedArticles",
        "description": "Returns a list of articles authored (created) by the currently authenticated user.\n\nRequires a valid Bearer token to identify the user.\n\nThis endpoint filters articles in the database by `ownerId` field, matching the authenticated user's ID.\n\n### Use cases:\n- Show a user their authored content\n- Allow user to manage/edit their own posts\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of created articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Created articles fetched successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Article"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users/save/{articleId}": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Save an article to the user's saved list",
        "operationId": "saveArticle",
        "description": "Saves an article to the authenticated user's saved list.\n\nRequires a valid Bearer token.\n\n### Behavior:\n- If the article does not exist, returns 404.\n- If the article is already saved, returns 409 Conflict.\n- On success, returns the updated list of saved article IDs.\n\n### Use cases:\n- Allow users to bookmark or collect articles\n- Enable frontend to indicate saved state per article\n",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the article to be saved"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Article saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Article saved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "64a1cbe2f4891c001e57a1f5"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Article not found"
          },
          "409": {
            "description": "Article already saved"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Remove an article from the user's saved list",
        "operationId": "removeSavedArticle",
        "description": "Removes an article from the authenticated user's saved articles list.\n\nRequires a valid Bearer token.\n\n### Behavior:\n- If the article is not in the user's saved list, returns 404.\n- On success, returns the updated saved articles list.\n\n### Use cases:\n- Allow users to un-save or remove bookmarked articles\n",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the article to be removed from saved list"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Article removed from saved list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Article removed from saved list"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "64a1cbe2f4891c001e57a1f5"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Article not found in saved list"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a list of all article authors (with optional filters and pagination)",
        "operationId": "getAllUsers",
        "description": "Returns a list of users who have authored articles on the site.\n\nSupports filtering, limiting, and optional pagination via query parameters.\n\n### Optional query parameters:\n- **filter=popular** — return users sorted by the number of created articles in descending order.\n- **limit=number** — limit the number of returned users (e.g. `limit=6`).\n  If pagination is not used, this limit still applies.\n- **page=number** — used for pagination (e.g. `page=2`).\n- **perPage=number** — number of users per page (default is `20`).\n\n### Example use cases:\n- Show top 6 authors: `/api/users?filter=popular&limit=6`\n- Paginate full author list: `/api/users?page=2&perPage=10`\n- Get all authors sorted by popularity (default limit 20): `/api/users?filter=popular`\n\nThis endpoint does **not** require authentication.\n",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string",
              "enum": [
                "popular"
              ]
            },
            "required": false,
            "description": "Sort users by number of articles in descending order."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "description": "Limit the number of users returned. Works with or without pagination."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "description": "Page number (for pagination)."
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "description": "Number of users per page (default is 20, for pagination)."
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Users fetched successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserPublic_User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 20
                        },
                        "totalItems": {
                          "type": "integer",
                          "example": 86
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 5
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by ID",
        "operationId": "getUserById",
        "description": "Retrieves public information of a specific user by their unique identifier.\n\n  This endpoint is typically used to:\n  - Display the author’s profile on an article page\n  - Fetch data to populate user profile modals or pages\n  - View a contributor’s details (e.g., in admin dashboards)\n\n  ### Behavior:\n  - Only returns non-sensitive fields (e.g., name, avatar, email)\n  - Returns a 404 error if the user does not exist\n  - The `id` parameter must be a valid MongoDB ObjectId\n\n  ### Example use cases:\n  - **/users/64b7dbf2eac4a914b7d6e217** → returns user details for that ID\n  - When viewing an article written by a specific user\n\n  > ⚠️ This endpoint does not require authentication, but only public data will be returned.\n",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the user",
            "schema": {
              "type": "string",
              "example": "6881563901add19ee16fcff2"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "User profile data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User fetched successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserPublic_User"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users/info": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update user's personal information",
        "operationId": "updateUserInfo",
        "description": "Allows the authenticated user to update their profile information such as name and email.\n\n### Requirements:\n- **Authentication**: Required via Bearer Token\n- Both `name` and `email` are optional, but at least one must be provided\n\n### Validation:\n- **Name**:\n  - Must be a string\n  - 2–32 characters\n- **Email**:\n  - Must be a valid email address\n  - Maximum 64 characters\n\n### Behavior:\n- If both fields are missing, validation middleware should reject the request\n- Updates only the fields provided in the request body\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 32,
                    "example": "Jane Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 64,
                    "example": "jane@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User info updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserPublic_User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/follow/{targetUserId}": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Follow another user",
        "operationId": "followUser",
        "description": "Authenticated user follows another user by their ID.\n\n❗ You cannot subscribe to yourself or re-subscribe to the same user.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "targetUserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to follow"
          }
        ],
        "responses": {
          "200": {
            "description": "Follow action completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Followed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          }
        }
      }
    },
    "/users/unfollow/{targetUserId}": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Unfollow a user",
        "operationId": "unfollowUser",
        "description": "Allows an authenticated user to unfollow another user by their ID.\n\n❗ If the user is not subscribed to the target, the request passes without error, just without changes.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "targetUserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to unfollow"
          }
        ],
        "responses": {
          "200": {
            "description": "Unfollow successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Unfollowed successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          }
        }
      }
    },
    "/users/followers/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get followers of a user",
        "operationId": "getFollowers",
        "description": "Returns a list of users who follow the specified user.\n\nUseful for an author's profile to show their audience.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user whose followers to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "List of followers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Followers list"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserPublic_User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          }
        }
      }
    },
    "/users/following/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get following list of a user",
        "operationId": "getFollowing",
        "description": "Returns a list of users whom the specified user is following.\n\nFor example, you can display who a user is reading on their profile.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user whose followings to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "List of followings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Followings list"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserPublic_User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound_NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Article": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a1cbe2f4891c001e57a1f5"
          },
          "img": {
            "type": "string",
            "example": "https://example.com/image.jpg",
            "description": "Image file (JPEG/PNG, max 1MB)"
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 48,
            "example": "The Benefits of Music Therapy"
          },
          "article": {
            "type": "string",
            "minLength": 100,
            "maxLength": 4000,
            "example": "Music therapy has shown benefits in reducing anxiety and improving mood."
          },
          "ownerId": {
            "type": "string",
            "example": "64a1cbe2f4891c001e57a1d1"
          },
          "ownerName": {
            "type": "string",
            "minLength": 4,
            "maxLength": 50,
            "example": "John Doe"
          },
          "rate": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "example": 4.5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-07-25T12:34:56.789Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-07-26T08:15:00.123Z"
          }
        }
      },
...
      "UserPublic_User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "64a1cbe2f4891c001e57a1d1"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "avatar": {
            "type": "string",
            "example": "https://res.cloudinary.com/demo/image/upload/v12345678/avatar.jpg",
            "nullable": true
          },
          "articlesAmount": {
            "type": "integer",
            "example": 0
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                }
              }
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Invalid or expired token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
...
      "NotFound_NotFound": {
        "description": "User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}